pyproject.toml
[project]
name = "brain-mcp"
version = "0.2.0"
description = "MCP server for the Brain Obsidian knowledge base"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.26.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[project.scripts]
brain-mcp = "brain_mcp.server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brain_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-anyio>=0.0.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
